+++ sqlobject/sqlbuilder.py
@@ -480,7 +480,8 @@ class Table(SQLExpression):
 
     def __getattr__(self, attr):
         if attr.startswith('__'):
-            raise AttributeError
+            raise AttributeError("Attribute '%s' is forbidden in '%s'" % (
+                attr, self.__class__.__name__))
         return self.FieldClass(self.tableName, attr)
 
     def __sqlrepr__(self, db):
@@ -502,7 +503,8 @@ class SQLObjectTable(Table):
 
     def __getattr__(self, attr):
         if attr.startswith('__'):
-            raise AttributeError
+            raise AttributeError("Attribute '%s' is forbidden in '%s'" % (
+                attr, self.__class__.__name__))
         if attr == 'id':
             return self._getattrFromID(attr)
         elif attr in self.soClass.sqlmeta.columns:
@@ -565,14 +567,16 @@ class TableSpace:
 
     def __getattr__(self, attr):
         if attr.startswith('__'):
-            raise AttributeError
+            raise AttributeError("Attribute '%s' is forbidden in '%s'" % (
+                attr, self.__class__.__name__))
         return self.TableClass(attr)
 
 
 class ConstantSpace:
     def __getattr__(self, attr):
         if attr.startswith('__'):
-            raise AttributeError
+            raise AttributeError("Attribute '%s' is forbidden in '%s'" % (
+                attr, self.__class__.__name__))
         return SQLConstant(attr)
 
 
@@ -628,7 +632,8 @@ class AliasTable(Table):
 
     def __getattr__(self, attr):
         if attr.startswith('__'):
-            raise AttributeError
+            raise AttributeError("Attribute '%s' is forbidden in '%s'" % (
+                attr, self.__class__.__name__))
         if self.table:
             attr = getattr(self.table.q, attr).fieldName
         return self.FieldClass(self.tableName, attr, self.alias, self)
@@ -638,15 +643,54 @@ class AliasTable(Table):
                              self.alias)
 
 
+class AliasSQLMeta():
+    def __init__(self, table, alias):
+        self.__table = table
+        self.__alias = alias
+
+    def __getattr__(self, attr):
+        if attr.startswith('__'):
+            raise AttributeError("Attribute '%s' is forbidden in '%s'" % (
+                attr, self.__class__.__name__))
+        table = self.__table
+        if (attr == "table"):
+            return '%s  %s' % (table.sqlmeta.table, self.__alias)
+        return getattr(table.sqlmeta, attr)
+
+
 class Alias(SQLExpression):
     def __init__(self, table, alias=None):
         self.q = AliasTable(table, alias)
+
+    def __getattr__(self, attr):
+        table = self.q.table
+        if (attr == "sqlmeta") and hasattr(table, "sqlmeta"):
+            alias = self.q.alias
+            return AliasSQLMeta(table, alias)
+        return getattr(table, attr)
 
     def __sqlrepr__(self, db):
         return sqlrepr(self.q, db)
 
     def components(self):
         return [self.q]
+
+    def select(self, clause=None, clauseTables=None,
+               orderBy=NoDefault, limit=None,
+               lazyColumns=False, reversed=False,
+               distinct=False, connection=None,
+               join=None, forUpdate=False):
+        # Import here to avoid circular import
+        from .sresults import SelectResults
+        return SelectResults(self, clause,
+                             clauseTables=clauseTables,
+                             orderBy=orderBy,
+                             limit=limit,
+                             lazyColumns=lazyColumns,
+                             reversed=reversed,
+                             distinct=distinct,
+                             connection=connection,
+                             join=join, forUpdate=forUpdate)
 
 
 class Union(SQLExpression):